cmd: Create webrtc test entrypoint with sub-roles
