+++ cmd/webrtc-load-tester/roles/streamer.go
@@ -1,0 +1,51 @@
+package roles
+
+import (
+	"flag"
+	"fmt"
+	"os"
+	"runtime"
+	"time"
+
+	"github.com/livepeer/stream-tester/internal/utils"
+	"github.com/livepeer/stream-tester/model"
+	"github.com/peterbourgon/ff/v2"
+)
+
+type streamerArguments struct {
+	Verbosity int
+	Version   bool
+
+	URL          string
+	TestDuration time.Duration
+}
+
+func Streamer() {
+	var cliFlags = streamerArguments{}
+
+	flag.Set("logtostderr", "true")
+	vFlag := flag.Lookup("v")
+
+	fs := flag.NewFlagSet("webrtc-load-tester", flag.ExitOnError)
+
+	fs.IntVar(&cliFlags.Verbosity, "v", 3, "Log verbosity.  {4|5|6}")
+	fs.BoolVar(&cliFlags.Version, "version", false, "Print out the version")
+
+	fs.DurationVar(&cliFlags.TestDuration, "test-dur", 0, "How long to run overall test")
+
+	_ = fs.String("config", "", "config file (optional)")
+
+	ff.Parse(fs, os.Args[1:],
+		ff.WithConfigFileFlag("config"),
+		ff.WithConfigFileParser(ff.PlainParser),
+		ff.WithEnvVarPrefix("LT_WEBRTC"),
+	)
+	flag.CommandLine.Parse(nil)
+	vFlag.Value.Set(fmt.Sprintf("%d", cliFlags.Verbosity))
+
+	hostName, _ := os.Hostname()
+	fmt.Println("WebRTC Load Tester streamer version: " + model.Version)
+	fmt.Printf("Compiler version: %s %s\n", runtime.Compiler, runtime.Version())
+	fmt.Printf("Hostname %s OS %s IPs %v\n", hostName, runtime.GOOS, utils.GetIPs())
+	fmt.Printf("Production: %v\n", model.Production)
+}
